+++ src/components/layouts/header.tsx
@@ -17,7 +17,7 @@ const Header = () => {
       <nav className="flex items-center gap-3">
         {currentUserId ? (
           <>
-            <Link href="/mypage">내 여행</Link>
+            <Link href="/my-plan">내 여행</Link>
             <Link href="/shared-plan">커뮤니티</Link>
             <MypageButton />
           </>
